entities: move debackbonified functions to alt modules
Inaugurating the pattern of leaving Backbone lib modules alone
and create new independent _alt modules for the needs of Svelte components,
with the major difference that they don't use Backbone Model and Collections
