+++ app/modules/entities/lib/types/author.js
@@ -1,6 +1,5 @@
 import { i18n } from 'modules/user/lib/i18n'
 import preq from 'lib/preq'
-import { attachEntities, getEntitiesByUris } from '../entities'
 
 export default function () {
   // Main property by which sub-entities are linked to this one
@@ -63,26 +62,4 @@ const getWorksUris = (works, seriesUris) => {
   return works
   .filter(workData => !seriesUris.includes(workData.serie))
   .map(getUri)
-}
-
-export const addAuthorWorks = async author => {
-  const { uri } = author
-  const { articles, series, works } = await preq.get(app.API.entities.authorWorks(uri))
-  const seriesUris = series.map(getUri)
-  const worksUris = getWorksUris(works, seriesUris)
-  const articlesUris = getWorksUris(articles, seriesUris)
-  await Promise.all([
-    attachEntities(author, 'articles', articlesUris),
-    attachEntities(author, 'series', seriesUris),
-    attachEntities(author, 'works', worksUris),
-  ])
-  return author
-}
-
-export const getAuthorWorks = async ({ uri }) => {
-  const { works } = await preq.get(app.API.entities.authorWorks(uri))
-  const worksUris = works.map(getUri)
-  const worksEntities = await getEntitiesByUris(worksUris)
-  // Filtering-out any non-work undetected by the SPARQL query
-  return worksEntities.filter(entity => entity.type === 'work')
 }
